[CAKER-2] Enable macOS target compatibility
- adjusted package platform list
- separated AppKit/UIKit specific implementation into `HapticPlayer` component, used by a common service abstraction
